azure/repind-cati-ss/catiApi/src/main/java/com/afklm/cati/common/spring/rest/resources/UserResource.java
package com.afklm.cati.common.spring.rest.resources;

import java.io.Serializable;

import javax.validation.constraints.NotNull;

import org.hibernate.validator.constraints.Email;
import org.hibernate.validator.constraints.Length;

public class UserResource extends CatiCommonResource implements Serializable {

    /**
     * 
     */
    private static final long serialVersionUID = 4634609266319780882L;
    
    
    @Email
    @NotNull
    private String mail;
    
    @Length(min = 2)
    @NotNull
    private String firstName;
    
    @NotNull
    @Length(min = 2)
    private String lastName;
    
    public String getMail() {
        return mail;
    }
    public void setMail(String mail) {
        this.mail = mail;
    }
    public String getFirstName() {
        return firstName;
    }
    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }
    public String getLastName() {
        return lastName;
    }
    public void setLastName(String lastName) {
        this.lastName = lastName;
    }
}
